+++ crates/coeus-cuda/src/backend/ops/impls/pool.rs
@@ -210,7 +210,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
         dilation: usize,
         output: &mut Self::DeviceBuffer<T>,
         output_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         self.cuda_max_pool1d(
             input,
             input_layout,
@@ -220,7 +220,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
             dilation,
             output,
             output_layout,
-        );
+        )
     }
 
     fn max_pool1d_backward(
@@ -235,7 +235,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
         dilation: usize,
         grad_input: &mut Self::DeviceBuffer<T>,
         grad_input_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         self.cuda_max_pool1d_backward(
             grad_out,
             grad_out_layout,
@@ -247,7 +247,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
             dilation,
             grad_input,
             grad_input_layout,
-        );
+        )
     }
 
     fn avg_pool1d(
@@ -260,7 +260,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
         dilation: usize,
         output: &mut Self::DeviceBuffer<T>,
         output_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         self.cuda_avg_pool1d(
             input,
             input_layout,
@@ -270,7 +270,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
             dilation,
             output,
             output_layout,
-        );
+        )
     }
 
     fn avg_pool1d_backward(
@@ -283,7 +283,7 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
         dilation: usize,
         grad_input: &mut Self::DeviceBuffer<T>,
         grad_input_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         self.cuda_avg_pool1d_backward(
             grad_out,
             grad_out_layout,
@@ -293,6 +293,6 @@ impl<T: CudaScalar + hephaestus_cuda::DialectScalar<hephaestus_cuda::CudaC>> coe
             dilation,
             grad_input,
             grad_input_layout,
-        );
+        )
     }
 }