+++ crates/coeus-ops/src/backend_ops/cpu_impl/impls/pool.rs
@@ -229,7 +229,7 @@ where
         dilation: usize,
         output: &mut Self::DeviceBuffer<T>,
         output_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         pool::max_pool1d(
             self,
             input,
@@ -241,6 +241,7 @@ where
             output,
             output_layout,
         );
+        Ok(())
     }
 
     #[inline]
@@ -256,7 +257,7 @@ where
         dilation: usize,
         grad_input: &mut Self::DeviceBuffer<T>,
         grad_input_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         pool::max_pool1d_backward(
             self,
             grad_out,
@@ -270,6 +271,7 @@ where
             grad_input,
             grad_input_layout,
         );
+        Ok(())
     }
 
     #[inline]
@@ -283,7 +285,7 @@ where
         dilation: usize,
         output: &mut Self::DeviceBuffer<T>,
         output_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         pool::avg_pool1d(
             self,
             input,
@@ -295,6 +297,7 @@ where
             output,
             output_layout,
         );
+        Ok(())
     }
 
     #[inline]
@@ -308,7 +311,7 @@ where
         dilation: usize,
         grad_input: &mut Self::DeviceBuffer<T>,
         grad_input_layout: &Layout,
-    ) {
+    ) -> Result<(), Self::Error> {
         pool::avg_pool1d_backward(
             self,
             grad_out,
@@ -320,5 +323,6 @@ where
             grad_input,
             grad_input_layout,
         );
+        Ok(())
     }
 }